+++ react/javascript/.storybook/webpack.config.js
@@ -14,7 +14,11 @@ module.exports = ({ config }) => {
         loader: require.resolve('raw-loader'),
       },
     ],
-  })
+  });
+  config.module.rules.push({
+    test: /\.scss$/,
+    use: ['style-loader', 'css-loader', 'sass-loader'],
+  });
   config.resolve.extensions.push('.ts', '.tsx')
   return config
 }